Use typedef long INT also works

+++ lapack_compatible_sources/include/config_dgeqp4.h
@@ -1,1 +1,1 @@
-typedef int INT;
+typedef long INT;
